+++ Source/Public/Get-PlexShare.ps1
@@ -1,5 +1,36 @@
 function Get-PlexShare
 {
+	<#
+		.SYNOPSIS
+			Gets a user and the share status of your libraries with them.
+
+		.DESCRIPTION
+			Gets a user and the share status of your libraries with them.
+
+		.PARAMETER Username
+			The username of the user to query share status.
+
+		.PARAMETER Email
+			The email address of the user to query share status.
+
+		.EXAMPLE
+			Get-PlexShare -Username "username"
+
+		.EXAMPLE
+			# Get share status for a single user:
+			Get-PlexShare -Username "username" | Select -ExpandProperty section
+
+		.EXAMPLE
+			# Get share status for all users:
+			Get-PlexUser | Select username | % { Get-PlexShare -Username $_.username }
+
+		.OUTPUTS
+			username allowSync section                                 invitedAt
+			-------- --------- -------                                 ---------
+			person1  1         {Section, Section, Section, Section...} 16/01/2022 19:01:39
+			person2  0         {Section, Section, Section, Section...} 08/01/2022 20:15:31
+	#>
+
 	[CmdletBinding(SupportsShouldProcess)]
 	param(
 		[Parameter(Mandatory = $false, ParameterSetName = 'username')]
@@ -75,7 +106,7 @@ function Get-PlexShare
 		[array]$Results = $Data.MediaContainer.SharedServer | Where-Object { $_."$($PsCmdlet.ParameterSetName)" -eq $($PSBoundParameters[$PsCmdlet.ParameterSetName]) }
 		if(!$Results)
 		{
-			Write-Verbose -Message "No results found for specified username"
+			Write-Verbose -Message "Function: $($MyInvocation.MyCommand): No results found for specified username"
 			return
 		}
 	}