+++ Source/Public/New-PlexPlaylist.ps1
@@ -105,7 +105,7 @@ function New-PlexPlaylist
 	#Region Make request
 	if($PSCmdlet.ShouldProcess($Item.title, "Add label '$Label'"))
 	{
-		Write-Verbose -Message "Adding label '$Label' to item '$($Item.title)'"
+		Write-Verbose -Message "Function: $($MyInvocation.MyCommand): Adding label '$Label' to item '$($Item.title)'"
 		try
 		{
 			$Data = Invoke-RestMethod -Uri $DataUri -Method POST